fix memset on x86-64 with a value of 0
the error is in the expression val & 0xff == 0 because in C == binds stronger
than & resulting in val & (0xff == 0) i.e. val & 0 i.e. 0. This error was
particularly bad when memset was called with a 16 byte aligned size and value 0
as it then zeroed register 0 (rax) which might be used for something else.