Optimize a jump-around-branch case
In my Emacs JIT I noticed that sometimes libjit would emit a
conditional jump to branch around an unconditional jump, like:

    7ffff7fe5160:	0f 85 05 00 00 00    	jne    0x7ffff7fe516b
    7ffff7fe5166:	e9 0e 00 00 00       	jmpq   0x7ffff7fe5179
    7ffff7fe516b:	41 bf 17 00 00 00    	mov    $0x17,%r15d

While this can be worked around by constructing the IL a bit
differently, I was curious about the difficulty of fixing this in the
libjit optimizer.

This patch notices the specific case where a conditional branch simply
jumps around an unconditional branch.  In this case, the condition is
inverted and the extra jump is eliminated.