make sure values don't change their address
the current implementations of incoming_frame_posn and outgoing_frame_posn
tend to break addressable variables, as they simply change a values
frame_offset.
For incoming_frame_posn this is fine, as it is only used at the start
of a function.
This commit changes the implementation of outgoing_frame_posn to store
the value at the target frame offset.